# openapi.yaml
openapi: 3.1.0
info:
  title: RiteSmith API
  version: 0.1.0
  summary: Capability-aware code and workflow generation engine.
  description: |
    RiteSmith is a capability-aware generation engine for safe, reusable automation artifacts.

    It can discover available tools/capabilities, search its memory/repository for reusable artifacts,
    generate Lua scripts, shell scripts, and Trama workflows, validate generated artifacts with
    guardrails, persist approved artifacts, and optionally delegate execution to external runtimes.

    RiteSmith is not just a storage service. It is the generator, reviewer, validator, catalog-aware
    planner, and artifact repository.

    Main responsibilities:

    - discover available tools and providers
    - search reusable scripts/workflows from memory
    - generate Lua scripts
    - generate shell scripts
    - generate Trama workflows
    - compose scripts and workflows into executable plans
    - validate artifacts against guardrails
    - persist generated artifacts with metadata and audit trail
    - delegate execution to external runtimes when allowed

    Security model:

    - generation is not execution
    - execution requires explicit policy and runtime permission
    - dangerous capabilities require approval/policy checks
    - artifacts should be versioned, auditable, and reproducible
servers:
  - url: http://localhost:8081
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /plans:
    post:
      tags: [Plans]
      summary: Create a generation plan from user intent
      operationId: createPlan
      description: |
        Main RiteSmith entrypoint.

        Given a user intent, RiteSmith searches memory, discovers capabilities, proposes a plan,
        generates or reuses artifacts, validates them, and optionally persists them.

        A plan may produce:
        - Lua scripts
        - shell scripts
        - Trama workflow definitions
        - composed execution plans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
            examples:
              bitcoin_drop_monitor:
                value:
                  intent: Notify me if Bitcoin drops 3% from its current price.
                  requested_artifact_types: [lua_script, trama_workflow]
                  mode: propose
                  constraints:
                    require_approval_for_execution: true
                    allow_shell: false
                    allow_network: true
                  context:
                    room_id: room_personal
      responses:
        "201":
          description: Plan created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /plans/{planId}:
    get:
      tags: [Plans]
      summary: Get plan details
      operationId: getPlan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      responses:
        "200":
          description: Plan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /plans/{planId}/approve:
    post:
      tags: [Plans]
      summary: Approve a proposed plan
      operationId: approvePlan
      description: |
        Approves a proposed plan. Approval may allow persistence, workflow creation,
        or execution depending on policy.
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: Plan approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /plans/{planId}/reject:
    post:
      tags: [Plans]
      summary: Reject a proposed plan
      operationId: rejectPlan
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionRequest"
      responses:
        "200":
          description: Plan rejected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /plans/{planId}/artifacts:
    post:
      tags: [Plans, Artifacts]
      summary: Persist generated artifacts from a plan
      operationId: persistPlanArtifacts
      parameters:
        - $ref: "#/components/parameters/PlanId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersistArtifactsRequest"
      responses:
        "201":
          description: Artifacts persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, expired, or revoked access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Operation is not allowed by policy or permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict with current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    ArtifactType:
      type: string
      enum:
        - lua_script
        - shell_script
        - trama_workflow
        - workflow_template
        - prompt_template
        - capability_adapter
        - plan
    ArtifactStatus:
      type: string
      enum:
        - draft
        - proposed
        - validated
        - approved
        - active
        - deprecated
        - rejected
        - archived
    PlanStatus:
      type: string
      enum:
        - draft
        - proposed
        - validating
        - blocked
        - approved
        - rejected
        - persisted
        - executing
        - completed
        - failed
        - cancelled
    GenerationMode:
      type: string
      enum:
        - propose
        - generate
        - validate_only
        - persist
        - execute
...
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
    PolicyDecisionValue:
      type: string
      enum: [allow, deny, require_approval]
...
    PlanConstraints:
      type: object
      properties:
        allow_shell:
          type: boolean
          default: false
        allow_network:
          type: boolean
          default: true
        allow_filesystem:
          type: boolean
          default: false
        allow_workflow_creation:
          type: boolean
          default: true
        require_approval_for_execution:
          type: boolean
          default: true
        max_runtime_seconds:
          type: integer
          minimum: 1
        polling_interval_seconds:
          type: integer
          minimum: 1
        allowed_capabilities:
          type: array
          items:
            type: string
        denied_capabilities:
          type: array
          items:
            type: string
      additionalProperties: true
    CreatePlanRequest:
      type: object
      required: [intent]
      properties:
        intent:
          type: string
          minLength: 1
        requested_artifact_types:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactType"
        mode:
          $ref: "#/components/schemas/GenerationMode"
        constraints:
          $ref: "#/components/schemas/PlanConstraints"
        context:
          type: object
          additionalProperties: true
        reuse_policy:
          type: string
          enum: [prefer_reuse, force_new, reuse_or_generate]
          default: prefer_reuse
      additionalProperties: false
    Plan:
      type: object
      required: [plan_id, status, intent, created_at]
      properties:
        plan_id:
          type: string
        status:
          $ref: "#/components/schemas/PlanStatus"
        intent:
          type: string
        summary:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PlanStep"
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        validations:
          type: array
          items:
            $ref: "#/components/schemas/ValidationResult"
        policy_decision:
          $ref: "#/components/schemas/PolicyDecision"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: false
    PlanStep:
      type: object
      required: [step_id, title, status]
      properties:
        step_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped, blocked]
        artifact_id:
          type: string
        capability_id:
          type: string
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
    ApprovalRequest:
      type: object
      properties:
        comment:
          type: string
      additionalProperties: false
    RejectionRequest:
      type: object
      properties:
        reason:
          type: string
      additionalProperties: false
    PersistArtifactsRequest:
      type: object
      properties:
        artifact_ids:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        tags:
          type: array
          items:
            type: string
      additionalProperties: false
...
    ValidationResult:
      type: object
      required: [valid, checks]
      properties:
        valid:
          type: boolean
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ValidationCheck"
        warnings:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            type: string
      additionalProperties: false
    ValidationCheck:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [passed, warning, failed, skipped]
        message:
          type: string
      additionalProperties: false
...
    Artifact:
      type: object
      required: [artifact_id, artifact_type, name, version, status, content, created_at]
      properties:
        artifact_id:
          type: string
        artifact_type:
          $ref: "#/components/schemas/ArtifactType"
        name:
          type: string
        description:
          type: string
        version:
          type: integer
          minimum: 1
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        content:
          type: string
        content_hash:
          type: string
        generated_by_plan_id:
          type: string
        source_artifact_ids:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        validation:
          $ref: "#/components/schemas/ValidationResult"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: false
    ArtifactListResponse:
      type: object
      required: [artifacts]
      properties:
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
      additionalProperties: false
...
    PolicyDecision:
      type: object
      required: [decision]
      properties:
        decision:
          $ref: "#/components/schemas/PolicyDecisionValue"
        reason:
          type: string
        required_approvals:
          type: array
          items:
            type: string
        blocked_capabilities:
          type: array
          items:
            type: string
      additionalProperties: false